requirements.txt
# psycopg2==2.9.10
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
dbt-core==1.7.18
dbt-postgres==1.7.18
# pre-commit==4.0.1
# autoflake==2.3.1
# black==24.10.0
# reorder-python-imports==3.14.0
# flake8==7.1.1
# flake8-annotations==3.1.1
# flake8-annotations-coverage==0.0.6
# flake8-bandit==4.1.1
# flake8-broken-line==1.0.0
# flake8-bugbear==24.10.31
# flake8-builtins==2.5.0
# flake8-commas==4.0.0
# flake8-comprehensions==3.16.0
# flake8-eradicate==1.5.0
# flake8-executable==2.1.3
# flake8-future-import==0.4.7
# flake8-pyi==24.9.0
# flake8-mutable==1.2.0
# flake8-string-format==0.3.0
# flake8-unused-arguments==0.0.13
# yesqa==1.5.0
# sqlfluff
# sqlfluff-templater-dbt